requirements.txt
asgiref==3.4.1
Django==3.2.6
django-nested-admin==3.3.3
pkg_resources==0.0.0
python-monkey-business==1.0.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
